+++ OpenUtau/Views/MainWindow.axaml.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Reactive;
 using System.Threading.Tasks;
 using Avalonia;
+using Avalonia.Collections;
 using Avalonia.Controls;
 using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Controls.Primitives;
@@ -89,9 +90,6 @@ namespace OpenUtau.App.Views {
 
         public void InitProject() {
             viewModel.InitProject(this);
-            if(viewModel.ShowWelcomePage) {
-                WelcomePage?.Show();
-            }
         }
 
         void OnEditTimeSignature(object sender, PointerPressedEventArgs args) {
@@ -202,6 +200,7 @@ namespace OpenUtau.App.Views {
             if (!DocManager.Inst.ChangesSaved && !await AskIfSaveAndContinue()) {
                 return;
             }
+            viewModel.Page = 1;
             viewModel.NewProject();
         }
 
@@ -222,6 +221,7 @@ namespace OpenUtau.App.Views {
             if (files == null || files.Length == 0) {
                 return;
             }
+            viewModel.Page = 1;
             try {
                 viewModel.OpenProject(files);
             } catch (Exception e) {
@@ -842,6 +842,7 @@ namespace OpenUtau.App.Views {
                 if (!DocManager.Inst.ChangesSaved && !await AskIfSaveAndContinue()) {
                     return;
                 }
+                viewModel.Page = 1;
                 try {
                     viewModel.OpenProject(new string[] { file });
                 } catch (Exception e) {
@@ -849,6 +850,7 @@ namespace OpenUtau.App.Views {
                     _ = await MessageBox.ShowError(this, new MessageCustomizableException($"Failed to open file {file}", $"<translate:errors.failed.openfile>: {file}", e));
                 }
             } else if (ext == ".mid" || ext == ".midi") {
+                viewModel.Page = 1;
                 try {
                     viewModel.ImportMidi(file);
                 } catch (Exception e) {
@@ -905,6 +907,7 @@ namespace OpenUtau.App.Views {
                     DependencyInstaller.Install(file);
                 }
             } else if (ext == ".mp3" || ext == ".wav" || ext == ".ogg" || ext == ".flac") {
+                viewModel.Page = 1;
                 try {
                     viewModel.ImportAudio(file);
                 } catch (Exception e) {
